serwo/examples/ImageProcessAz/cnn/requirements.txt
boto3
Pillow
torch
torchvision